requirements.txt
pillow
pyowm
